Modificacoes pos novos enactors

+++ src/java/br/ufrn/controlers/AtualizarWidgets.java
@@ -14,28 +14,28 @@ public interface AtualizarWidgets {
 
     public void AddAtualizadoresWidgets(Integer area, AtualizarWidgets atualizarWidgetsArea);
     //=====PollutionWidget=====
-    public void atualizarIndicePoluicao(Integer area, int indexPolluition);
+    public void atualizarIndicePoluicao(Integer area, Integer indexPolluition);
 
-    public void atualizarFluxoDeVeiculos(Integer area, int vehicleFlow);
+    public void atualizarFluxoDeVeiculos(Integer area, Integer vehicleFlow);
 
     //este metodo deve atualizar o PollutionWidget e o TemperatureWidget
-    public void atualizarVelociadeDoVento(Integer area, int windSpeed);
+    public void atualizarVelociadeDoVento(Integer area, Integer windSpeed);
 
     public void atualizardirecaoDoVento(Integer area, String directionWind);
 
-    public void atualizarProbalidadeChuva(Integer area, int willRain);
+    public void atualizarProbalidadeChuva(Integer area, Integer willRain);
 
-    public void atualizarQuantidadeRuido(Integer area, int noise);
+    public void atualizarQuantidadeRuido(Integer area, Integer noise);
 
     //======TemperatureWidget=====
-    public void atualizarTemeperatura(Integer area, float temperature);
+    public void atualizarTemeperatura(Integer area, Integer temperature);
 
-    public void atualizarHumidade(Integer area, int humidity);
+    public void atualizarHumidade(Integer area, Integer humidity);
 
     public void atualizarIncidenciaDeIncendio(Integer area, boolean isFire);
 
     //=====DumpWidget===
-    public void atualizarQuantidadeLixo(Integer area, int trash);
+    public void atualizarQuantidadeLixo(Integer area, Integer trash);
 
     public void atualizarAgenteProximo(Integer area, String agentNearName);
 }
